update api
추가 api
1. 뉴스 전체조회
2. 뉴스 전체 카테고리 조회
3. 사용자 이메일 --> 사용자 아이디조회 --> 컨텐츠 추천

남은일 : 뉴스 i.d로 뉴스조회 추가

+++ src/main/java/org/example/mlooops/entity/UserInterestEntity.java
@@ -21,9 +21,9 @@ public class UserInterestEntity {
     private int categoryId;
 
     @Column(name = "category_Double", nullable = false)
-    private double categoryDouble;
+    private float categoryDouble;
 
-    public UserInterestEntity(int userId, int categoryId, double categoryDouble) {
+    public UserInterestEntity(int userId, int categoryId, float categoryDouble) {
         this.userId = userId;
         this.categoryId = categoryId;
         this.categoryDouble = categoryDouble;
